requirements.txt
aiogram
cryptocompare
python-decouple
peewee
psycopg2-binary
peewee-migrate
aiosqlite
asyncpg
aioredis==1.3.1
loguru
aiogram-datepicker
pendulum
pyTelegramBotAPI
pytz
alembic
aiohttp_middlewares